+++ tools/phpstan/castor.php
@@ -7,13 +7,12 @@ namespace Flow\Tools;
 use Castor\Attribute\AsTask;
 
 use function Castor\run;
-use function dirname;
 
 #[AsTask(description: 'Execute PHPStan analysis', aliases: ['phpstan'])]
 function phpstan(): int
 {
     return run(
-        [__DIR__ . '/vendor/bin/phpstan', '--configuration=' . dirname(__DIR__, 2) . '/phpstan.neon'],
+        [__DIR__ . '/vendor/bin/phpstan', '--configuration=' . __DIR__ . '/phpstan.neon'],
         allowFailure: true,
     )->getExitCode();
 }